typos and added RoundSkeleton in it's own file
